requirements.txt
annotated-types==0.7.0
anyio==4.7.0
azure-core==1.32.0
black==24.10.0
boto3==1.35.86
botocore==1.35.86
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
eval_type_backport==0.2.0
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.0
idna==3.10
jiter==0.8.2
jmespath==1.0.1
json5==0.10.0
jsonpath-python==1.0.6
mistralai==1.2.5
mypy-extensions==1.0.0
openai==1.58.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3